The same code for powers of three.
Begin by converting every integer, k, to a binary bit-string, then finding the base3
value of that string of 1's and 0's. That's the sum of the values of
3^i, for all values of i where the bit is set.

All the rest of the math can stay the same. (k).bit_count() is still the
size of the subset, and j & k still tells whether the subsets are
disjoint.